requirements.txt
# ── Scraping ──────────────────────────────────────────────
httpx==0.27.2
tenacity==9.0.0
selectolax==0.3.21
beautifulsoup4==4.12.3
pdfplumber==0.11.4
camelot-py[cv]==0.11.0
wbgapi==1.0.12

# ── Pipeline & Data ──────────────────────────────────────
polars==1.17.1
pandas==2.2.3
pandera==0.20.4
pyyaml==6.0.2

# ── Analytics & Notebooks ────────────────────────────────
jupyter==1.1.1
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1
geopandas==1.0.1
mapclassify==2.8.1
scikit-learn==1.5.2
shapely==2.0.6

# ── Dashboard ────────────────────────────────────────────
streamlit==1.40.2
streamlit-folium==0.23.1
folium==0.18.0

# ── Report Generation ────────────────────────────────────
nbconvert==7.16.4
weasyprint==63.1

# ── Testing ──────────────────────────────────────────────
pytest==8.3.4
pytest-httpx==0.34.0
